requirements.txt
dash>=2.18
dash-bootstrap-components>=1.6
flask>=3.0
flask-login>=0.6
sqlalchemy>=2.0
pandas>=2.0
openpyxl>=3.1
waitress>=3.0
python-dateutil>=2.9
